move floating menu widget and controller to a separate file
