+++ lib/src/deriv_chart/deriv_chart.dart
@@ -12,7 +12,6 @@ import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/chart_serie
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/markers/marker_series.dart';
 import 'package:deriv_chart/src/deriv_chart/chart/data_visualization/models/chart_object.dart';
 import 'package:deriv_chart/src/deriv_chart/drawing_tool_chart/drawing_tools.dart';
-import 'package:deriv_chart/src/deriv_chart/interactive_layer/interactive_layer.dart';
 import 'package:deriv_chart/src/misc/callbacks.dart';
 import 'package:deriv_chart/src/misc/chart_controller.dart';
 import 'package:deriv_chart/src/models/chart_axis_config.dart';
@@ -28,6 +27,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'interactive_layer/interactive_layer_behaviours/interactive_layer_behaviour.dart';
 import 'interactive_layer/interactive_layer_behaviours/interactive_layer_desktop_behaviour.dart';
 import 'interactive_layer/interactive_layer_behaviours/interactive_layer_mobile_behaviour.dart';
+import 'interactive_layer/interactive_layer_controller.dart';
 
 /// A wrapper around the [Chart] which handles adding indicators to the chart.
 class DerivChart extends StatefulWidget {
