+++ lib/src/deriv_chart/interactive_layer/widgets/selected_drawing_floating_menu.dart
@@ -1,0 +1,53 @@
+import 'package:deriv_chart/src/add_ons/drawing_tools_ui/drawing_tool_config.dart';
+import 'package:deriv_chart/src/deriv_chart/interactive_layer/interactable_drawings/interactable_drawing.dart';
+import 'package:flutter/material.dart';
+
+import '../interactive_layer_controller.dart';
+
+/// A floating menu that appears when a drawing is selected.
+class SelectedDrawingFloatingMenu extends StatefulWidget {
+  /// Creates a floating menu for the selected drawing.
+  const SelectedDrawingFloatingMenu({
+    required this.drawing,
+    required this.controller,
+    super.key,
+  });
+
+  /// The drawing that is currently selected.
+  final InteractableDrawing<DrawingToolConfig> drawing;
+
+  /// The controller for the interactive layer.
+  final InteractiveLayerController controller;
+
+  @override
+  State<SelectedDrawingFloatingMenu> createState() =>
+      _SelectedDrawingFloatingMenuState();
+}
+
+class _SelectedDrawingFloatingMenuState
+    extends State<SelectedDrawingFloatingMenu> {
+  @override
+  Widget build(BuildContext context) => Positioned(
+        left: widget.controller.floatingMenuPosition.dx,
+        top: widget.controller.floatingMenuPosition.dy,
+        child: GestureDetector(
+          behavior: HitTestBehavior.opaque,
+          onPanUpdate: (details) {
+            widget.controller.floatingMenuPosition += details.delta;
+            setState(() => {});
+          },
+          child: Container(
+            decoration: BoxDecoration(
+              color: Colors.red,
+              borderRadius: BorderRadius.circular(8),
+            ),
+            child: Row(
+              children: [
+                const Icon(Icons.drag_handle, color: Colors.white),
+                Text(widget.drawing.runtimeType.toString())
+              ],
+            ),
+          ),
+        ),
+      );
+}
